pyproject.toml
[tool.poetry]
name = "chat-extension"
version = "0.1.0"
description = "Interatct with any website through a chatbot"
authors = ["frederic.legrand"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
scrapy = "^2.11.2"
sentence-transformers = "^3.1.1"
qdrant-client = "^1.11.3"
mistralai = "^1.1.0"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
